requirements.txt
torch==2.4.1
torchvision==0.19.1
numpy==1.26.4
scikit-learn==1.4.2
pillow==10.3.0
matplotlib==3.8.4
tqdm==4.66.4
umap-learn==0.5.5
pytest==8.2.2
